==> note/2_Not_hot_songs.ipynb <==
# Lab_Not_hot_songs


from bs4 import BeautifulSoup
import pandas as pd 
import numpy as np
import requests

nothot100 = pd.read_csv("../data/raw_data/spotify_millsongdata.csv")

nothot100.head()

nothot100.shape

nothot100 = nothot100[["artist","song"]]
nothot100.head()

nothot100 = nothot100.rename(columns={'artist': 'Artist', 'song': 'Title'})
nothot100.head()

nothot100.to_csv('../data/nothot100.csv', index=False)
