==> note/3_spotify-api-song_15000_-1.ipynb <==
# Spotify_API

### This note get the more information of hot 100 and non hot 100 from Spotify 

from bs4 import BeautifulSoup
import pandas as pd 
import numpy as np
import requests

#load not_hot100
nothot100 = pd.read_csv("../data/nothot100.csv")
nothot100.head(2)

len(nothot100)

from credentials import *

import spotipy
import json
from spotipy.oauth2 import SpotifyClientCredentials
import os

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=client_id,
                                                          client_secret=client_secret))

### get the ID from Spotify

def search_song(title, artist):
    
    query = "tracks: "+title+" artist: "+artist
    #print(query)

    try:
        result = sp.search(query, limit=1)
        song_id = result['tracks']['items'][0]['id']
    except:
        song_id = "not_found"

    return song_id  

search_song("dancing queen", "abba")

### choose the sample

random_seed = 42
df2 = nothot100.sample(n=15000, random_state=random_seed)
df2.isnull().sum()

df2 = df2.drop_duplicates()
df2.shape

df2 = df2.dropna()
len(df2)

## apply get_audio_features
def get_audio_features(list_of_song_ids):
    
    chunks = [list_of_song_ids[i:i + 50] for i in range(0, len(list_of_song_ids), 50)]
    
    # Create a dictionary to store the audio features
    audio_features_dict = {
        'id': [],
        'danceability': [],
        'energy': [],
        'key': [],
        'loudness': [],
        'mode': [],
        'speechiness': [],
        'acousticness': [],
        'instrumentalness': [],
        'liveness': [],
        'valence': [],
        'tempo': [],
        'type': [],
        'uri': [],
        'track_href': [],
        'analysis_url': [],
        'duration_ms': [],
        'time_signature': []
    }
    
    # Retrieve the audio features for each chunk
    for chunk in chunks:
        audio_features = sp.audio_features(chunk)
        
        # Append the audio features to the dictionary
        for feature in audio_features:
            if feature:
                for key in audio_features_dict.keys():
                    audio_features_dict[key].append(feature[key])
        
        # Wait 20 seconds before next API call
        time.sleep(10)

    # Create a DataFrame from the dictionary
    df = pd.DataFrame(audio_features_dict)

    return df

### df2 

df2_1 = df2.iloc[:5000,:]
df2_2 = df2.iloc[5000:10000,:]
df2_3 = df2.iloc[10000:,:]

# create the chunks
chunks_df2_1 = [df2_1[i:i + 50] for i in range(0, len(df2_1), 50)]
chunks_df2_2 = [df2_2[i:i + 50] for i in range(0, len(df2_2), 50)]
chunks_df2_3 = [df2_3[i:i + 50] for i in range(0, len(df2_3), 50)]

df2.to_csv('../data/1choose_15000songs.csv', index=False)

df2.shape

#df2
import time
song_ids_df2_1 = []
count = 0
for chunk in chunks_df2_1:
    for index, row in chunk.iterrows():
        song_id = search_song(row["Title"], row["Artist"])
        if song_id == 'not_found':
            print("yes")
        song_ids_df2_1.append(song_id)
    print(count, end="")
    count = count + 1
    time.sleep(10)
print('finished')

#save
txt_file_path = '../data/list_data_songID_5000_1.txt'

# Write the list elements to the text file
with open(txt_file_path, 'w') as file:
    for item in song_ids_df2_1:
        file.write(str(item) + '\n')

display(len(song_ids_df2_1))
set_list_1 = set(song_ids_df2_1)
display(len(set_list_1))

#df2
#df2
import time
song_ids_df2_2 = []
count = 0
for chunk in chunks_df2_2:
    for index, row in chunk.iterrows():
        song_id = search_song(row["Title"], row["Artist"])
        if song_id == 'not_found':
            print("yes")
        song_ids_df2_2.append(song_id)
    print(count, end="")
    count = count + 1
    time.sleep(10)
print('finished')

#save
txt_file_path = '../data/list_data_songID_5000_2.txt'

# Write the list elements to the text file
with open(txt_file_path, 'w') as file:
    for item in song_ids_df2_2:
        file.write(str(item) + '\n')

#df2
import time
song_ids_df2_3 = []
count = 0
for chunk in chunks_df2_3:
    for index, row in chunk.iterrows():
        song_id = search_song(row["Title"], row["Artist"])
        if song_id == 'not_found':
            print("yes")
        song_ids_df2_3.append(song_id)
    print(count, end="")
    count = count + 1
    time.sleep(10)
print('finished')

txt_file_path = '../data/list_data_songID_5000_3.txt'

# Write the list elements to the text file
with open(txt_file_path, 'w') as file:
    for item in song_ids_df2_3:
        file.write(str(item) + '\n')

song_ids_df2 = song_ids_df2_1 + song_ids_df2_2 + song_ids_df2_3
len(song_ids_df2)

#to get the index which has 'not_found'
index_list = []
for i in range(0,15000):
    if song_ids_df2[i] == 'not_found':
        index_list.append(i)
index_list    

song_ids_df3 = song_ids_df2.copy()
#remove 'not_found' values 
song_ids_df3 = [x for x in song_ids_df3 if x != 'not_found']
len(song_ids_df3)

txt_file_path = '../data/song_ids_15000.txt'

# Write the list elements to the text file
with open(txt_file_path, 'w') as file:
    for item in song_ids_df3:
        file.write(str(item) + '\n')

# remove specific rows based on the given indices
df8 = df2.copy()
df9 = df8.reset_index(drop=True)
df10 = df9.drop(index_list)
len(df10)


id_1 = song_ids_df3[:5000]
id_2 = song_ids_df3[5000 : 10000]
id_3 = song_ids_df3[10000 :]

## apply get_audio_features
def get_audio_features1(list_of_song_ids):
    count = 0
    chunks = [list_of_song_ids[i:i + 50] for i in range(0, len(list_of_song_ids), 50)]
    
    # Create a dictionary to store the audio features
    audio_features_dict = {
        'id': [],
        'danceability': [],
        'energy': [],
        'key': [],
        'loudness': [],
        'mode': [],
        'speechiness': [],
        'acousticness': [],
        'instrumentalness': [],
        'liveness': [],
        'valence': [],
        'tempo': [],
        'type': [],
        'uri': [],
        'track_href': [],
        'analysis_url': [],
        'duration_ms': [],
        'time_signature': []
    }
    
    # Retrieve the audio features for each chunk
    for chunk in chunks:
        audio_features = sp.audio_features(chunk)
        
        # Append the audio features to the dictionary
        for feature in audio_features:
            if feature:
                for key in audio_features_dict.keys():
                    audio_features_dict[key].append(feature[key])
        
        # Wait 15 seconds before next API call
        print(count,end='')
        count = count + 1
        time.sleep(30)

    # Create a DataFrame from the dictionary
    df = pd.DataFrame(audio_features_dict)

    return df

df11_1 = get_audio_features1(id_1)

df11_1.to_csv('../data/only__audio_featured_not_hot100_1.csv', index=False)

df11_2 = get_audio_features1(id_2)

df11_2.to_csv('../data/only__audio_featured_not_hot100_2.csv', index=False)

df11_3 = get_audio_features1(id_3)

df11_3.to_csv('../data/only__audio_featured_not_hot100_3.csv', index=False)

# Concatenate the three DataFrames vertically
concat15000 = pd.concat([df11_1, df11_2, df11_3])
concat15000

concat15000.isnull().sum()

concat15000.to_csv('../data/only__audio_featured_not_hot100.csv', index=False)

len(df10)

#### continue the next notebook: 3_spotify-api-song_15000_-2
