==> note/3_spotify-api-song_15000_-2.ipynb <==
### This note is the sequeal of '3_spotify-api-song_15000'

from bs4 import BeautifulSoup
import pandas as pd 
import numpy as np
import requests

name = pd.read_csv("../data/1choose_15000songs.csv")
name.head(2)


file_path = '../data/list_data_songID_5000_1.txt'

with open(file_path, 'r') as file:
    list_1 = [line.strip() for line in file]


file_path2 = '../data/list_data_songID_5000_2.txt'

with open(file_path2, 'r') as file:
    list_2 = [line.strip() for line in file]

file_path3 = '../data/list_data_songID_5000_3.txt'

with open(file_path3, 'r') as file:
    list_3 = [line.strip() for line in file]

id_list = list_1 + list_2 + list_3
len(id_list)

#to get the index which has 'not_found'
index_list = []
for i in range(0,15000):
    if id_list[i] == 'not_found':
        index_list.append(i)
index_list    

song_ids_df3 = id_list.copy()
#remove 'not_found' values 
song_ids_df3 = [x for x in song_ids_df3 if x != 'not_found']
len(song_ids_df3)

# remove specific rows based on the given indices
name1 = name.copy()
name1 = name1.reset_index(drop=True)
name = name.drop(index_list)
len(name)


# create a datafram of 'ID'

id_data = pd.DataFrame({'id': song_ids_df3})
id_data.head()

id_data1 = id_data.copy()
id_data2 = id_data.copy()
# Create a list of indexes that need to be removed from df2 (indexes of duplicates in df1)
indexes_to_remove = id_data1.index[id_data1.duplicated()].tolist()
display(len(id_data2.duplicated()))

# Remove rows from df2 with the same indexes as in indexes_to_remove
name2 = name.copy()
name2 = name2.reset_index(drop=True)
name2.drop(indexes_to_remove, inplace=True)

display(len(name2))

name2

id_data2.drop(indexes_to_remove, inplace=True)

display(len(id_data2))

id_data2 = id_data2.reset_index(drop=True)
name2 = name2.reset_index(drop=True)

concatenated_df = pd.concat([name2, id_data2], axis=1)
concatenated_df

audio_feature = pd.read_csv("../data/only__audio_featured_not_hot100.csv")
audio_feature

audio_feature = audio_feature.drop_duplicates()
audio_feature

#join 

# Perform the inner join based on the 'ID' column
inner_join_df = pd.merge(concatenated_df, audio_feature, on='id', how='inner')
inner_join_df

inner_join_df.to_csv('../data/9138_audio_featured_songs.csv', index=False)
