==> note/5._confirmation.ipynb <==
import pandas as pd
songs = pd.read_csv('../data/song_recommender.csv')

songs

df = songs
import pandas as pd

# Assuming 'df' is your DataFrame
# Check if 'id' column contains the value '5u2vOSftv32iQpSZvP6ugV'
value_to_find = '5u2vOSftv32iQpSZvP6ugV'
is_value_present = value_to_find in df['id'].values

# Find row index(es) where the value exists
if is_value_present:
    row_indices = df.index[df['id'] == value_to_find].tolist()
    print(f"The value '{value_to_find}' is present in the 'id' column.")
    print(f"The row index(es) where the value is present: {row_indices}")
else:
    print(f"The value '{value_to_find}' is not present in the 'id' column.")


df.iloc[813,:]
